Allow constructing only a subset of R columns
The interface for set_R_correlated() and set_R_correlated() is extended
with two optional parameters j_start and j_end. This is to facilitate
cases where there is insufficient memory to store R on a single node,
and also to facilitate parallelization over multiple GPUs.

In the new code, the RNG argument is used to generate seeds for new
random number generators, one for each column. In this way, the columns
of R can be generated in serial or parallel, with exactly the same
realizations of random numbers.